subiendo

+++ src/services/doctor/delete/doctorDelete.service.ts
@@ -1,4 +1,4 @@
-import { Inject, Injectable } from '@nestjs/common';
+import { Inject, Injectable, NotFoundException } from '@nestjs/common';
 import { DoctorResponseDto } from 'src/domain/entities/doctor/dto/response/doctorResponse.dto';
 import { IDoctorRepository } from 'src/domain/interfaces/infrastructure/doctor/IDoctor.repository';
 import { IDoctorDeleteService } from 'src/domain/interfaces/services/doctor/delete/IDoctorDeleteService';
@@ -11,13 +11,18 @@ export class DoctorDeleteService implements IDoctorDeleteService {
   ) {}
 
   /**
-   * delete doctor
+   * Eliminar un doctor
    * @param _id
    * @returns
    */
   async delete(_id: string): Promise<DoctorResponseDto> {
     try {
+      const searchDoctor = await this._doctorRepository.findOne({ _id });
+
+      if (!searchDoctor) throw new NotFoundException('Este doctor no existe');
+
       const deleteDoctor = await this._doctorRepository.delete({ _id });
+
       return deleteDoctor;
     } catch (error) {
       throw new Error(error);
